+++ src/konpy/answer.clj
@@ -313,14 +313,28 @@
 (defn- inner-link [s]
   [:a.underline {:href (str "#" s)} s])
 
+(defn- week-num [eid]
+  (-> (db/q '[:find ?week ?num
+              :keys week num
+              :in $ ?eid
+              :where
+              [?e :week ?week]
+              [?e :num ?num]
+              [(= ?e ?eid)]] eid)
+      first))
+
+; (week-num 7745)
+
 (defn answers-others
   [{{:keys [e]} :path-params}]
-  (let [answers (->> (db/q q-answers-others (parse-long e))
+  (let [eid (parse-long e)
+        answers (->> (db/q q-answers-others eid)
                      (sort-by :updated)
-                     reverse)]
+                     reverse)
+        {:keys [week num]} (week-num eid)]
     (page
      [:div {:class "mx-4 my-2"}
-      [:div {:class "text-2xl"} "現在までの回答数(人数): "
+      [:div {:class "text-2xl"} (str week "-" num " 現在までの回答数(人数): ")
        (count answers) " (" (-> (map :author answers) set count) ")"]
       [:div.py-2 (interpose \space (mapv #(inner-link (:author %)) answers))]
       (for [a answers]